requirements.txt
aws.cdk.core
aws_cdk.cloudformation_include